chore(server): use decimal

+++ server/db/migrate/20190327074009_create_cell_outputs.rb
@@ -1,7 +1,7 @@
 class CreateCellOutputs < ActiveRecord::Migration[5.2]
   def change
     create_table :cell_outputs do |t|
-      t.decimal :capacity, precision: 32, scale: 2
+      t.decimal :capacity, precision: 64, scale: 2
       t.binary :data
       t.belongs_to :ckb_transaction, index: true
